refactor: extract supervisor tools
